Improve editor command line arguments: (#34)
Instead of assuming the command line arguments are fit for opening in a
particular line and column, we now translate based on the command line.

The only proper way would be to include information on how each editor
needs its arguments.

If the editor is unknown, we simply use the editor parameter with the
filename, which is the only likely thing to at least work:

    $ENV{EDITOR} file.pl

If we were to try to also introduce the line and column numbers, it is
more likely to fail than succeed because each editor parses command line
differently.

A problem that still needs to be addressed is the usage of
$App::Critique::CONFIG{CRITIQUE_EDITOR} as the same value as
$App::Critique::COFNIG{EDITOR} as the same value as $ENV{EDITOR}.

These are different and should not be mixed together. It would likely
require some more subtle logic in Critique.pm, hopefully removing the
BEGIN {} block at the process. :)

This small bit includes a test. It is not numbered because I don't see
the point of numbering tests.